refactor: remove outdated `no_std` part of zenoh_util
